backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
pydantic-settings==2.2.1
pymongo==4.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
celery==5.4.0
redis==5.0.4
httpx==0.27.0
websockets==12.0
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.1.0
shap==0.45.1
joblib==1.4.2
rasterio==1.3.10
shapely==2.0.4
Pillow==10.3.0
boto3==1.34.0
pytest==8.2.0
pytest-asyncio==0.23.0
